feat: default pageEnd to 1 and make it inclusive instead of exclusive with pageStart
The reason we default pageEnd to 1 is because that's basically all you
are able to get currently because pageEnd's implementation has changed
on YouTube. You can no longer use it in the query string on the /results
endpoint to get more videos -- the way you get more results is by
scrolling down on the page which triggers a kind of continuation POST
requests with json data of more results. Will have to do investigate
this further.